move helpers in utils file
